# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Система управления учета оборудования и мебели",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/cabinets": {
            "get": {
                "tags": [
                    "Cabinets"
                ],
                "summary": "Получить список всех кабинетов",
                "description": "Display a listing of the resource.",
                "operationId": "f13450e6142746037bb952b26d4cb2b2",
                "responses": {
                    "200": {
                        "description": "Список кабинетов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/CabinetIndexResource"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cabinets"
                ],
                "summary": "Создать новый кабинет",
                "description": "Store a newly created resource in storage.",
                "operationId": "8ab811dddb177c25085fed41762051de",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCabinetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Кабинет создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CabinetShowResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cabinets/{id}": {
            "get": {
                "tags": [
                    "Cabinets"
                ],
                "summary": "Получить кабинет",
                "description": "Display the specified resource.",
                "operationId": "14d6656346529841789dbc512f2c7bf4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID кабинета",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Кабинет",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/CabinetShowResource"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cabinets/{cabinet}": {
            "delete": {
                "tags": [
                    "Cabinets"
                ],
                "summary": "Удалить кабинет",
                "description": "Remove the specified resource from storage.",
                "operationId": "ed55c4186ff7baaf452bc4dab2b65ad6",
                "parameters": [
                    {
                        "name": "cabinet",
                        "in": "path",
                        "description": "ID кабинета",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Кабинет удален"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Cabinets"
                ],
                "summary": "Обновить кабинет",
                "description": "Update the specified resource in storage.",
                "operationId": "14d9f7c4c8f96935519ec7312012f08e",
                "parameters": [
                    {
                        "name": "cabinet",
                        "in": "path",
                        "description": "ID кабинета",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCabinetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Кабинет изменен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CabinetShowResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/cabinets/buildings/{building}/floor/{floor}": {
            "get": {
                "tags": [
                    "Cabinets"
                ],
                "summary": "Получить список кабинетов по этажам площадки",
                "description": "Display a listing of the resource.",
                "operationId": "6c16cc482843dabb88e1cebfc88f9a50",
                "parameters": [
                    {
                        "name": "building",
                        "in": "path",
                        "description": "ID площадки",
                        "required": true,
                        "example": 1
                    },
                    {
                        "name": "floor",
                        "in": "path",
                        "description": "Номер этажа",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список кабинетов по этажам площадки",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/CabinetIndexResource"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreCabinetRequest": {
                "title": "Запрос на создание кабинета",
                "description": "Данные для создания нового кабинета",
                "required": [
                    "cabinet_type_id",
                    "building_id",
                    "title",
                    "cabinet_number",
                    "floor"
                ],
                "properties": {
                    "cabinet_type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "building_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Лаборатория информационных технологий"
                    },
                    "cabinet_number": {
                        "type": "integer",
                        "example": 321
                    },
                    "floor": {
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
            "UpdateCabinetRequest": {
                "title": "Запрос на изменение кабинета",
                "description": "Данные для обновления кабинета",
                "required": [
                    "cabinet_type_id",
                    "building_id",
                    "title",
                    "cabinet_number",
                    "floor"
                ],
                "properties": {
                    "cabinet_type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "building_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Лаборатория информационных технологий"
                    },
                    "cabinet_number": {
                        "type": "integer",
                        "example": 321
                    },
                    "floor": {
                        "type": "integer",
                        "example": 3
                    }
                },
                "type": "object"
            },
...
            "CabinetIndexResource": {
                "title": "Ресурс списка кабинетов",
                "description": "Формат данных кабинета при получении списка",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "cabinet_type": {
                                    "type": "string",
                                    "example": "Лаборатория"
                                },
                                "building": {
                                    "type": "string",
                                    "example": "Анадырский"
                                },
                                "title": {
                                    "type": "string",
                                    "example": "Лаборатория информационных технологий"
                                },
                                "cabinet_number": {
                                    "type": "integer",
                                    "example": 321
                                },
                                "floor": {
                                    "type": "integer",
                                    "example": 3
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "CabinetShowResource": {
                "title": "Ресурс кабинета",
                "description": "Формат данных кабинета",
                "properties": {
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "cabinet_type": {
                                "type": "string",
                                "example": "Лаборатория"
                            },
                            "building": {
                                "type": "string",
                                "example": "Анадырский"
                            },
                            "title": {
                                "type": "string",
                                "example": "Лаборатория информационных технологий"
                            },
                            "cabinet_number": {
                                "type": "integer",
                                "example": 321
                            },
                            "floor": {
                                "type": "integer",
                                "example": 3
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}